requirements.txt
PyYAML==6.0.1
sympy~=1.12
graphviz~=0.20.1